moved enums from DAO to API module
