+++ sm-api/src/main/java/cz/fi/muni/pa165/soccermanager/api/enums/Footed.java
@@ -1,0 +1,5 @@
+package cz.fi.muni.pa165.soccermanager.api.enums;
+
+public enum Footed {
+	RIGHT, LEFT, BOTH
+}
