+++ sm-api/src/main/java/cz/fi/muni/pa165/soccermanager/api/enums/Position.java
@@ -1,4 +1,4 @@
-package cz.fi.muni.pa165.soccermanager.data.enums;
+package cz.fi.muni.pa165.soccermanager.api.enums;
 
 public enum Position {
 	DEFFENSE, OFFENSE, MIDFIELD
