--- sm-dao/src/main/java/cz/fi/muni/pa165/soccermanager/data/enums/Footed.java
@@ -1,5 +1,0 @@
-package cz.fi.muni.pa165.soccermanager.data.enums;
-
-public enum Footed {
-	RIGHT, LEFT, BOTH
-}
